host/requirements.txt
pycryptodome==3.20.0
scapy==2.5.0
websockets==12.0
